+++ src/components/organisms/article/carousel/CarouselContainer/index.stories.tsx
