+++ src/components/organisms/article/carousel/PickupArticleCarousel/index.spec.tsx
