+++ src/components/organisms/article/carousel/PickupArticleCarousel/index.tsx
