+++ src/components/pages/Home/index.tsx
@@ -1,6 +1,6 @@
 import { HtmlHeadBase } from "@/components/functions/meta";
 import HomeLayout from "@/components/layouts/HomeLayout";
-import { ArticleSWRContainer } from "@/components/organisms/ArticleSWRContainer";
+import { ArticleSWRContainer } from "@/components/organisms/article";
 import type { TArticleListResponse, TCategory, TConfig, TPickup, TRankedArticle } from "@/types";
 
 export type HomeProps = {
