+++ src/components/pages/articles/categories/Category/index.tsx
@@ -1,6 +1,6 @@
 import { HtmlHeadBase } from "@/components/functions/meta";
 import DefaultLayout from "@/components/layouts/DefaultLayout";
-import { ArticleSWRContainer } from "@/components/organisms/ArticleSWRContainer";
+import { ArticleSWRContainer } from "@/components/organisms/article";
 import type { TArticleListResponse, TCategory, TConfig, TPickup, TRankedArticle } from "@/types";
 import { formatPageTitle } from "@/utils/formatter";
 import { getBackLinks, urlTable } from "@/utils/paths/url";
